+++ app/controllers/application_controller.rb
@@ -5,21 +5,27 @@ class ApplicationController < ActionController::Base
   skip_before_action :verify_authenticity_token, if: :devise_login?
   before_action :configure_permitted_parameters, if: :devise_controller?
 
-  def after_sign_in_path_for(resource)
-    session[:previous_internal_url] || session[:previous_external_url] || dashboard_path
-  end
-
-  def after_sign_out_path_for(resource_or_scope)
-    session[:previous_external_url] || root_path
-  end
+  # Concerns
+  include Forwardable
 
   private
+
+  def internal_controllers
+    {
+      admin: [:dashboard],
+      categories: [:edit, :reorder],
+      documents: [:index, :show, :new, :edit],
+      folders: [:index, :show, :new, :edit],
+      internal: [],
+      sites: [:index, :show, :new, :edit, :recruiting_centers, :coordinating_centers]
+    }
+  end
 
   def configure_permitted_parameters
     devise_parameter_sanitizer.permit(
       :sign_up,
       keys: [
-        :username, :full_name, :email, :password, :emails_enabled # :password_confirmation
+        :username, :full_name, :email, :password, :emails_enabled
       ]
     )
   end